requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv
sentence-transformers
streamlit
pypdf
requests
datasets
beautifulsoup4